move terminate and logs to backend topic and add aliases, also improve output for terminate command
